--- Open-Editor/ayarlar.py
@@ -1,60 +1,0 @@
-
-
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-
-class Ui_MainWindow(object):
-    def setupUi(self, MainWindow):
-        MainWindow.setObjectName("MainWindow")
-        MainWindow.resize(382, 292)
-        self.centralwidget = QtWidgets.QWidget(MainWindow)
-        self.centralwidget.setObjectName("centralwidget")
-        self.pushButton = QtWidgets.QPushButton(self.centralwidget)
-        self.pushButton.setGeometry(QtCore.QRect(0, 220, 381, 51))
-        self.pushButton.setObjectName("pushButton")
-        self.pushButton.clicked.connect(self.font_ayar)
-
-        self.label = QtWidgets.QLabel(self.centralwidget)
-        self.label.setGeometry(QtCore.QRect(0, 0, 141, 21))
-        self.label.setObjectName("label")
-        self.label_2 = QtWidgets.QLabel(self.centralwidget)
-        self.label_2.setGeometry(QtCore.QRect(0, 60, 181, 16))
-        self.label_2.setObjectName("label_2")
-        self.textEdit = QtWidgets.QTextEdit(self.centralwidget)
-        self.textEdit.setGeometry(QtCore.QRect(10, 20, 261, 31))
-        self.textEdit.setObjectName("textEdit")
-        self.textEdit_2 = QtWidgets.QTextEdit(self.centralwidget)
-        self.textEdit_2.setGeometry(QtCore.QRect(10, 80, 261, 31))
-        self.textEdit_2.setObjectName("textEdit_2")
-        MainWindow.setCentralWidget(self.centralwidget)
-        self.statusbar = QtWidgets.QStatusBar(MainWindow)
-        self.statusbar.setObjectName("statusbar")
-        MainWindow.setStatusBar(self.statusbar)
-
-        self.retranslateUi(MainWindow)
-        QtCore.QMetaObject.connectSlotsByName(MainWindow)
-
-    def retranslateUi(self, MainWindow):
-        _translate = QtCore.QCoreApplication.translate
-        MainWindow.setWindowTitle(_translate("MainWindow", "Ayarlar"))
-        self.pushButton.setText(_translate("MainWindow", "Onayla"))
-        self.label.setText(_translate("MainWindow", "Font Adı:"))
-        self.label_2.setText(_translate("MainWindow", "Font Boyu:"))
-    def font_ayar(self):
-        dosya=open("ayarlar.txt","w")
-        font_ad="FONT_ADI="+self.textEdit.toPlainText()+"\n"
-        font_boy="FONT_BOYU="+self.textEdit_2.toPlainText()+"\n"
-        dosya.write(font_ad)
-        dosya.write(font_boy)
-        print("işlem tamamlandı")
-
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    MainWindow = QtWidgets.QMainWindow()
-    ui = Ui_MainWindow()
-    ui.setupUi(MainWindow)
-    MainWindow.show()
-    sys.exit(app.exec_())
