--- Open-Editor/versiyon.txt
@@ -1,1 +1,0 @@
-versiyon=1.6.3
