+++ ayarlar.txt
@@ -1,0 +1,2 @@
+FONT_ADI=Times
+FONT_BOYU=12
